MAINT: add boxcox from xsf and remove cython
