+++ scipy/special/_special_ufuncs.cpp
@@ -11,6 +11,7 @@
 #include <xsf/bessel.h>
 #include <xsf/beta.h>
 #include <xsf/binom.h>
+#include <xsf/boxcox.h>
 #include <xsf/cdflib.h>
 #include <xsf/convex_analysis.h>
 #include <xsf/digamma.h>
@@ -73,6 +74,8 @@ extern const char *besselpoly_doc;
 extern const char *beta_doc;
 extern const char *betaln_doc;
 extern const char *binom_doc;
+extern const char *boxcox_doc;
+extern const char *boxcox1p_doc;
 extern const char *cbrt_doc;
 extern const char *cosdg_doc;
 extern const char *cosm1_doc;
@@ -123,6 +126,8 @@ extern const char *i0_doc;
 extern const char *i0e_doc;
 extern const char *i1_doc;
 extern const char *i1e_doc;
+extern const char *inv_boxcox_doc;
+extern const char *inv_boxcox1p_doc;
 extern const char *iv_doc;
 extern const char *iv_ratio_doc;
 extern const char *iv_ratio_c_doc;
@@ -334,6 +339,16 @@ _special_ufuncs_module_exec(PyObject *module)
         {static_cast<xsf::numpy::ff_f>(xsf::rel_entr), static_cast<xsf::numpy::dd_d>(xsf::rel_entr)}, "rel_entr",
         rel_entr_doc);
     PyModule_AddObjectRef(module, "rel_entr", rel_entr);
+
+    PyObject *boxcox = xsf::numpy::ufunc(
+        {static_cast<xsf::numpy::ff_f>(xsf::boxcox), static_cast<xsf::numpy::dd_d>(xsf::boxcox)}, "boxcox",
+        boxcox_doc);
+    PyModule_AddObjectRef(module, "boxcox", boxcox);
+
+    PyObject *boxcox1p = xsf::numpy::ufunc(
+        {static_cast<xsf::numpy::ff_f>(xsf::boxcox1p), static_cast<xsf::numpy::dd_d>(xsf::boxcox1p)}, "boxcox1p",
+        boxcox1p_doc);
+    PyModule_AddObjectRef(module, "boxcox1p", boxcox1p);
 
     PyObject *airy =
         xsf::numpy::ufunc({static_cast<xsf::numpy::f_ffff>(xsf::airy), static_cast<xsf::numpy::d_dddd>(xsf::airy),
@@ -637,6 +652,16 @@ _special_ufuncs_module_exec(PyObject *module)
         {static_cast<xsf::numpy::f_f>(xsf::cyl_bessel_i1e), static_cast<xsf::numpy::d_d>(xsf::cyl_bessel_i1e)}, "i1e",
         i1e_doc);
     PyModule_AddObjectRef(module, "i1e", i1e);
+
+    PyObject *inv_boxcox = xsf::numpy::ufunc(
+        {static_cast<xsf::numpy::ff_f>(xsf::inv_boxcox), static_cast<xsf::numpy::dd_d>(xsf::inv_boxcox)},
+        "inv_boxcox", inv_boxcox_doc);
+    PyModule_AddObjectRef(module, "inv_boxcox", inv_boxcox);
+
+    PyObject *inv_boxcox1p = xsf::numpy::ufunc(
+        {static_cast<xsf::numpy::ff_f>(xsf::inv_boxcox1p), static_cast<xsf::numpy::dd_d>(xsf::inv_boxcox1p)},
+        "inv_boxcox1p", inv_boxcox1p_doc);
+    PyModule_AddObjectRef(module, "inv_boxcox1p", inv_boxcox1p);
 
     PyObject *iv = xsf::numpy::ufunc(
         {static_cast<xsf::numpy::ff_f>(xsf::cyl_bessel_i), static_cast<xsf::numpy::dd_d>(xsf::cyl_bessel_i),
